fix/socket: fix delay calculation
Fixed addition overflow. Honestly, I have no idea about average_delay_base role in here. But the point is that while it's calculated, its value is wrapped. Math is complex here, I don't understand it, but it seems like the rest of the code expects it to be wrapped, etc.