fix/socket: don't send Fin twice
Mainstream uTP implementation, libutp, makes sure Fin packet is not sent twice. That's not the case with our implementation.
In our case, if we explicitly shutdown socket and then close it, Fin packet will be sent twice. Which is *probably* not what is expected.
I stress *probably* here, cause it's undocumented in the protocol specs.