feat/socket: implement graceful connection shutdown
When connection has received Fin, it has sent it's own Fin and it was acked by remote peer, connection is considered to be gracefully shutdown.
In such case it's state is `Closed`.
There's a new method `UtpStream::finalize()`, which creates a future that waits for notification when connection state gets `Closed`.